Struct pallet_nodle_staking::pallet::Pallet

pub struct Pallet<T>(_);

Implementations

impl<T: Config> Pallet<T>

pub fn set_invulnerables(
origin: OriginFor<T>,
invulnerables: Vec<T::AccountId>
) -> DispatchResultWithPostInfo

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

pub fn set_total_validator_per_round(
origin: OriginFor<T>,
new: u32
) -> DispatchResultWithPostInfo

Set the total number of validator selected per round

changes are not applied until the start of the next round

pub fn validator_join_pool(
origin: OriginFor<T>,
bond: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Join the set of validators pool

pub fn validator_exit_pool(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Request to exit the validators pool. If successful, the account is immediately removed from the validator pool to prevent selection as a validator, but unbonding is executed with a delay of BondedDuration rounds.

pub fn validator_bond_more(
origin: OriginFor<T>,
more: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Bond more for validator

pub fn validator_bond_less(
origin: OriginFor<T>,
less: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Bond less for validator

pub fn nominator_nominate(
origin: OriginFor<T>,
validator: T::AccountId,
amount: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

If caller is not a nominator, then join the set of nominators If caller is a nominator, then makes nomination to change their nomination state

pub fn nominator_denominate(
origin: OriginFor<T>,
validator: T::AccountId
) -> DispatchResultWithPostInfo

Revoke an existing nomination

pub fn nominator_denominate_all(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Quit the set of nominators and, by implication, revoke all ongoing nominations

pub fn nominator_bond_more(
origin: OriginFor<T>,
validator: T::AccountId,
more: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Bond more for nominators with respect to a specific validator

pub fn nominator_bond_less(
origin: OriginFor<T>,
validator: T::AccountId,
less: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Bond less for nominators with respect to a specific nominated validator

pub fn withdraw_unbonded(origin: OriginFor<T>) -> DispatchResultWithPostInfo

pub fn withdraw_staking_rewards(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

pub fn slash_cancel_deferred(
origin: OriginFor<T>,
session_idx: SessionIndex,
controllers: Vec<T::AccountId>
) -> DispatchResultWithPostInfo

Cancel enactment of a deferred slash.

Can be called by the T::SlashCancelOrigin.

Parameters: session index and validator list of the slashes for that session to kill.

impl<T: Config> Pallet<T>

pub fn update_validators_pool(
validator: T::AccountId,
total: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance
)

pub fn remove_from_validators_pool(validator: T::AccountId)

impl<T: Config> Pallet<T>

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

impl<T: Config> Pallet<T>

pub fn total(
) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance

Total capital locked by this staking pallet

impl<T: Config> Pallet<T>

pub fn active_session() -> SessionIndex

Current session index

impl<T: Config> Pallet<T>

pub fn validator_fee() -> Perbill

Commission percent taken off of rewards for all validators

impl<T: Config> Pallet<T>

pub fn validator_state<KArg>(
k: KArg
) -> Option<Validator<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>> where
KArg: EncodeLike<T::AccountId>,

Get validator state associated with an account if account is collating else None

impl<T: Config> Pallet<T>

pub fn nominator_state<KArg>(
k: KArg
) -> Option<Nominator<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>> where
KArg: EncodeLike<T::AccountId>,

Get nominator state associated with an account if account is nominating else None

impl<T: Config> Pallet<T>

pub fn total_selected() -> u32

The total validators selected every round

impl<T: Config> Pallet<T>

pub fn selected_validators() -> Vec<T::AccountId>

The validators selected for the current round

impl<T: Config> Pallet<T>

pub fn validator_pool(
) -> OrderedSet<Bond<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>>

The pool of validator validators, each with their total backing stake

impl<T: Config> Pallet<T>

pub fn exit_queue() -> OrderedSet<Bond<T::AccountId, SessionIndex>>

A queue of validators awaiting exit BondedDuration delay after request

impl<T: Config> Pallet<T>

pub fn at_stake<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorSnapshot<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance> where
KArg1: EncodeLike<SessionIndex>,
KArg2: EncodeLike<T::AccountId>,

Snapshot of validator nomination stake at the start of the round

impl<T: Config> Pallet<T>

pub fn staked<KArg>(
k: KArg
) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance where
KArg: EncodeLike<SessionIndex>,

Total backing stake for selected validators in the round

impl<T: Config> Pallet<T>

pub fn session_accumulated_balance<KArg>(
k: KArg
) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance where
KArg: EncodeLike<SessionIndex>,

Accumulated balances for the last Session Round

impl<T: Config> Pallet<T>

pub fn session_validator_reward<KArg>(
k: KArg
) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance where
KArg: EncodeLike<SessionIndex>,

Validator reward for the Session

impl<T: Config> Pallet<T>

pub fn points<KArg>(k: KArg) -> u32 where
KArg: EncodeLike<SessionIndex>,

Total points awarded to validator for block production in the round

impl<T: Config> Pallet<T>

pub fn awarded_pts<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> u32 where
KArg1: EncodeLike<SessionIndex>,
KArg2: EncodeLike<T::AccountId>,

Points for each validator per round

impl<T: Config> Pallet<T>

pub fn stake_rewards<KArg>(
k: KArg
) -> Vec<UnlockChunk<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>> where
KArg: EncodeLike<T::AccountId>,

stakers nodle rewards per session

impl<T: Config> Pallet<T>

pub fn slash_reward_proportion() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

impl<T: Config> Pallet<T>

pub fn slashing_spans<KArg>(k: KArg) -> Option<SlashingSpans> where
KArg: EncodeLike<T::AccountId>,

Snapshot of validator slash state

impl<T: Config> Pallet<T>

pub fn span_slash<KArg>(
k: KArg
) -> SpanRecord<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance> where
KArg: EncodeLike<(T::AccountId, u32)>,

Snapshot of validator slash state

impl<T: Config> Pallet<T>

pub fn validator_slash_in_session<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Option<(Perbill, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance)> where
KArg1: EncodeLike<SessionIndex>,
KArg2: EncodeLike<T::AccountId>,

All slashing events on validators, mapped by session to the highest slash proportion and slash value of the session.

impl<T: Config> Pallet<T>

pub fn nominator_slash_in_session<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Option<<<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance> where
KArg1: EncodeLike<SessionIndex>,
KArg2: EncodeLike<T::AccountId>,

All slashing events on nominators, mapped by session to the highest slash value of the session.

impl<T: Config> Pallet<T>

pub fn unapplied_slashes<KArg>(
k: KArg
) -> Vec<UnappliedSlash<T::AccountId, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance>> where
KArg: EncodeLike<SessionIndex>,

All unapplied slashes that are queued for later.

impl<T: Config> Pallet<T>

pub fn bonded_sessions() -> Vec<SessionIndex>

A mapping of still-bonded sessions