Trait pallet_nodle_staking::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type BondedDuration: Get<SessionIndex>;
    type SlashDeferDuration: Get<SessionIndex>;
    type MinSelectedValidators: Get<u32>;
    type MaxNominatorsPerValidator: Get<u32>;
    type MaxValidatorPerNominator: Get<u32>;
    type DefaultValidatorFee: Get<Perbill>;
    type DefaultSlashRewardProportion: Get<Perbill>;
    type DefaultSlashRewardFraction: Get<Perbill>;
    type DefaultStakingMaxValidators: Get<u32>;
    type DefaultStakingMinStakeSessionSelection: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type DefaultStakingMinValidatorBond: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type DefaultStakingMinNominationChillThreshold: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type DefaultStakingMinNominatorTotalBond: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type RewardRemainder: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type SessionInterface: SessionInterface<Self::AccountId>;
    type ValidatorRegistration: ValidatorRegistration<Self::AccountId>;
    type PalletId: Get<ModuleId>;
    type StakingLockId: Get<LockIdentifier>;
    type MaxChunkUnlock: Get<usize>;
    type CancelOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;

}

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

Handler for the unbalanced reduction when slashing a staker.

type BondedDuration: Get<SessionIndex>

Number of sessions that staked fund remain bonded for

type SlashDeferDuration: Get<SessionIndex>

Number of sessions that slashes are deferred by, after computation.

type MinSelectedValidators: Get<u32>

Minimum number of selected validators every round

type MaxNominatorsPerValidator: Get<u32>

Maximum nominators per validator

type MaxValidatorPerNominator: Get<u32>

Maximum validators per nominator

type DefaultValidatorFee: Get<Perbill>

Fee due to validators, set at genesis

type DefaultSlashRewardProportion: Get<Perbill>

Default Slash reward propostion, set at genesis

type DefaultSlashRewardFraction: Get<Perbill>

The proportion of the slashing reward to be paid out on the first slashing detection.

type DefaultStakingMaxValidators: Get<u32>

Maximum validators allowed to join the pool.

type DefaultStakingMinStakeSessionSelection: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Minimum stake required for any account to be in SelectedCandidates for the session

type DefaultStakingMinValidatorBond: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Minimum stake required for any account to be a validator candidate

type DefaultStakingMinNominationChillThreshold: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Minimum stake for any registered on-chain account to nominate

type DefaultStakingMinNominatorTotalBond: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Minimum stake for any registered on-chain account to become a nominator

type RewardRemainder: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

Tokens have been minted and are unused for validator-reward. See Era payout.

type SessionInterface: SessionInterface<Self::AccountId>

Interface for interacting with a session module.

type ValidatorRegistration: ValidatorRegistration<Self::AccountId>

Validate a user is registered

type PalletId: Get<ModuleId>

This pallet’s module id. Used to derivate a dedicated account id to store session rewards for validators and nominators in.

type StakingLockId: Get<LockIdentifier>

staking pallet Lock Identifier used for set_lock()

type MaxChunkUnlock: Get<usize>

Max number of unbond request supported by queue

type CancelOrigin: EnsureOrigin<Self::Origin>

The origin which can cancel a deferred slash. Root can always do this.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.